move cli flag generation to runtime
