+++ lib/index.js
@@ -31,6 +31,7 @@ const exportPlugins = (obj, mappings) => {
 };
 
 exportPlugins(module.exports, {
+	cli: () => require("./cli"),
 	AutomaticPrefetchPlugin: () => require("./AutomaticPrefetchPlugin"),
 	BannerPlugin: () => require("./BannerPlugin"),
 	Cache: () => require("./Cache"),
