+++ test/Cli.test.js
@@ -1,0 +1,7 @@
+const webpack = require("../");
+
+describe("Cli", () => {
+	it("should generate the correct cli flags", () => {
+		expect(webpack.cli.getFlags()).toMatchSnapshot();
+	});
+});
